Breakout setup, measure import, and package import steps from ecm_prep.main() into ECMPrepHelper.
